+++ CHANGELOG.md
@@ -54,6 +54,7 @@ branch is the latest unreleased version.
 - Fix Envelope field being printed as Time field when ColorSequenceKeypoint or
   NumberSequenceKeypoint is converted to a string.
 - Fix bounds error when calling ColorSequence.new or NumberSequence.new.
+- Fix ColorSequence.new and NumberSequence.new not verifying order of keypoints.
 
 See a [comparison with the previous version][cmp-imperative] for a thorough list of changes.